Ignore some SslContextBuilderTests on certain JDKs
Motivation:
The tests testUnsupportedPrivateKeyFailsFastForServer and testUnsupportedPrivateKeyFailsFastForClient are failing on RedHat/Fedora distributions of JDK 11.

The test try to assert that a certificate using an exotic curve is rejected by BoringSSL, and throws an SSLException when building the private key.

However, on RedHat JDK 11, this curve - 1.3.132.0.39 - is not in the JDK CurveDB, and so we get an IllegalArgumentException when creating the key manager.
The CurveDB on this distribution has a pretty limited set of curves available, so it might not be possible to find one that is accepted by the key manager, but rejected by the private key build.

Modification:
Make the tests look out for the IllegalArgumentException from the CurveDB, and if observed, abort the test as if by an assumption violation.

Result:
The tests will now be ignored, rather than fail the build, when running on JDKs that do not know about the curve used in the test.