Add a `fast` maven profile
Motivation:
To assist IDEs and other tools, it is often required to make `install` builds locally.
Since this can block work, it's desirable for these builds to be as fast as possible.

Modification:
Add a profile that skips many checks, to speed up the build.

Result:
It is now possible to make a clean build with `mvn clean install -Pfast -T1C` in (on my machine) under two minutes.