Fix ByteBufUtil.indexOf(buf, buf)
Motivation:
The ByteBufUtil.indexOf(buf, buf) method has historically been returning absolute indexes in the haystack.
With the introduction of the Two-Way algorithm, it inadvertently got changed to return offsets relative to the haystacks reader index.
That's a compatibility breaking change.

Modification:
Restore the old behavior of returning absolute indexes.

Result:
Compatibility is restored.
Fixes #11963

Note that the impact of this bug has been relatively minor inside Netty itself, since the produced index was only used for the error message when Http2ConnectionHandler failed to detect "HTTP/1" in the client preface.