ByteToMessageDecoder must  not call read() in channelReadComplete(...), if no channelRead(...) was fired. (#11966)
Motivation:

When ByteToMessageDecoder added in channelRead(), the following channelReadComplete() will lead ByteToMessageDecoder to call a read(),
which is not expected.

Modifications:

Add a field variable tracking the state of if ByteToMessageDecoder.channelRead() is fired after handler added. And if it was NOT fired, then the next channelReadComplete() would not call read()

Result:

Fixes https://github.com/netty/netty/issues/11965

Co-authored-by: Norman Maurer <norman_maurer@apple.com>