Try to not recycle objects back to terminated threads
Motivation:
The Recycler uses thread-local caches, and if objects are shared across threads, they will return to the thread-local pool they were allocated from.
However, if that thread has terminated, then it will have no use for any returning objects and the return process can be skipped.

Modification:
In the associated `onRemoval` for the thread-local, mark the local pool as dead, by setting the `pooledHandles` reference to `null`.
Returning objects to their original pools is then guarded by a null-check.
To make this thread-safe, the `pooledHandles` reference changes from `final` to `volatile`.
With a newer JVM baseline, we could have made the field plain, and null it out with an opaque store, because the `handle.toAvailable()` call otherwise gives us the compiler barriers we need to make our null-check thread-safe.

Result:
Recycler-able objects can be garbage collected sooner after the termination of the allocating thread.

Fixes #11864