Rename RecvByteBufAllocator to RecvBufferAllocator
Motivation:
Using "Buffer" in the name is more general.
The relevant ChannelOption is already using "buf".
In a future PR, I want to make it possible for these to allocate either new- or old-api buffers.

Modification:
Rename RecvByteBufAllocator to RecvBufferAllocator.
Also rename sub-classes, and associated methods, and parameters, and local variables and fields where necessary.

Result:
A more general name for the receive buffer allocator.