Add RecvBufferAllocator API for allocating Buffer instances
Motivation:
In order for transports to be able to read into new-api Buffers, we need the RecvBufferAllocator to be able to allocate such buffers.

Modification:
Add an allocate overload to RecvBufferAllocator.Handle that take a BufferAllocator and return a Buffer.
Update all implementing classes to support this new API, in addition to the existing ByteBuf API.

Result:
Our RecvBufferAllocator implementations can now allocate Buffers, in addition to ByteBufs.