Use BeforeAll instead of static block to load Brotli4j library (#12029)
Motivation:
As of today, Brotli4j does not support macOS aarch64. Currently, in `BrotliEncoderTest`, we use the static block to load the Brotli4j library. However, this could lead to failure when performing this action on macOS aarch64.

The `DisabledIf` loads the `BrotliEncoderTest` class to see if the platform is supported or not. Loading this class triggers the static block which eventually leads to `ExceptionInInitializerError`.

Modification:
Use `BeforeAll` instead of static block.

Result:
Fixes #12024