Make it possible to move read-/write-offset from buffer components
Motivation:
Component iteration is the only way that we can access any underlying native pointers for a buffer.
We need to do this when we make native IO calls with a buffer.
If our plan was to just collect the pointers and sizes in iovecs, then there would not have been a problem.
But if we want to make our IO calls immediately, from within the iteration, then we will need a way to update the offsets.

Modification:
Add a pair of addBytesRead and addBytesWritten methods to ReadableComponent and WritableComponent, respectively.

Result:
It will now be possible to make native IO calls from within component iteration *and* update the buffer offsets, without using a capturing lambda.
This potentially saves some allocation overhead on every IO call.