ByteBuffer from component iteration must keep unsafe memory alive
Motivation:
It is possible to use forEachReadable or forEachWritable to create a ByteBuffer that is backed by native memory from a Buffer.
If we are using Unsafe, then we cannot rely on the underlying ByteBuffer implementation to keep the native memory alive for as long as it is accessible.
We must guard against the situation where forEachReadable/Writable is used on an UnsafeBuffer, then a readable/writeableBuffer is created, and this buffer is siphoned out of the iteration, and then kept alive for longer than the UnsafeBuffer.
We can only keep the UnsafeMemory alive, though.
We cannot guard against use-after-free, because we don't have a way to nerf a ByteBuffer once it has been released.

Modification:
Change the PlatformDependent.directBuffer method to take an attachment object, and pass it to the DirectByteBuffer constructor.
Then use this from the UnsafeBuffer *Component implementation to ensure that the underlying UnsafeMemory is kept alive and strongly reachable by any ByteBuffer instances we create.

Result:
Reduce the possibility of causing JVM crash through misuse of the ability to get ByteBuffers out of an UnsafeBuffer.