Avoid masking the cause of availability failures for the Mac DNS provider
Motivation:
If the MacOSDnsServerAddressStreamProvider is unavailable, we'd the root cause of it to show up.

Modification:
Move the initialisation of the `MacOSDnsServerAddressStreamProvider.currentMappings` field to the constructor.
Object initialisers run before the constructor does, so since `MacOSDnsServerAddressStreamProvider.retrieveCurrentMappings` relies on the native code, it would throw an `UnsatisfiedLinkError` before `ensureAvailability` would get to run.
This way, it was masking the root cause of the unavailability.
Now, the `currentMappings` is initialised *after* we've called `ensureAvailability`, so we don't lose the root cause if we cannot use native DNS on Mac.

Result:
Easier to debug problems with initialising native MacOS DNS.