Add Buffer methods to read/write on NIO channels
Motivation:
The NIO transport needs to perform IO on NIO channels using Buffer instances, but NIO needs ByteBuffer instances.
Bridging this gap in the most efficient way possible depends on the Buffer implementation.
Therefor, Buffer needs methods to do IO on NIO channels.

Modification:
Add readIntoChannelWrite and writeFromChannelRead methods to Buffer.
The names may look awkward, but that is a consequence of trying to avoid confusion, when we are doing reads from the perspective of the buffer, but writes from the perspective of the channel, and vice versa.
The DefaultCompositeBuffer additionally has vectored implementations for Scatter/Gather channels.

Result:
It'll be possible to implement Buffer support in the NIO transport, in a more efficient way.