Expose OpenSSL security level in the SSL handler
Motivation:

From version 1.1.0, OpenSSL provides a "security level" configuration which specifies which minimum security parameters are allowed for connexions (SSL version, ciphers, signature algorithms, and so on).
Setting a value less secure than the platform's default one can be useful, for instance when wanting to connect to legacy servers.

Modifications:

Allow the user to set OpenSSL security level parameter as an SSL context option.

Result:

The user can specify a value for the OpenSSL security level, as an SSL context option.