Implement first-class support for the new Buffers in transports
Motivation:
Our transports should be able to directly read and write buffers with the new API, without going through a ByteBuf intermediary object.

Modification:
Implement reading and writing of Buffers, both for streaming channels, and for datagrams.

Result:
It is now in principle possible to use the new buffer API end-to-end.