Fix certificate compression test
Motivation:
Our `Future.syn()` method no longer directly throws the failure cause exception (in violation with its statically declared exceptions) but instead wraps the cause in a CompletionException.
The test was written for 4.1, and did not anticipate this.

Modification:
Make the test expect the CompletionException, and instead perhap its assert on the associated cause.

Result:
Tests pass again on the main branch.