Avoid reading uninitialised memory when making domain socket address
Motivation:
Stack variables contain uninitialised memory, unless explicitly initialised.
When we use recvfrom on a domain socket, it may not fill in any NULL terminator for the sockaddr_un->sun_path, since it technically provided the address length in the addrlen inout-parameter.
Thus, if we wish to later use strlen in createDomainDatagramSocketAddress to determine the length of the sockaddr_un->sun_path, we have to initialise the sockaddr to zero, so the NUL byte is already in place.
We have to do this on every iteration since we don't know the state of the memory after a failed or interrupted call.

Modification:
Always zero out the data allocated to the socket address, prior to calling recvfrom on domain sockets.

Result:
We no longer observe the occasional domain socket address that has some extra data in the end of its name.