HTTP/2 header validation must reject duplicate pseudo-headers
Motivation:
Duplicate HTTP/2 pseudo-headers are either client bugs (best case) or an indication of a request smuggling attack (worst case).
Header validation should reject this with a protocol error.
Doing so is also in line with the current HTTP/2 spec draft, which says:

> The same pseudo-header field name MUST NOT appear more than once in a field block. A field block for an HTTP request or response that contains a repeated pseudo-header field name MUST be treated as malformed (Section 8.1.1).

See https://httpwg.org/http2-spec/draft-ietf-httpbis-http2bis.html#name-http-control-data

Modification:
If header validation is enabled, and we parsed a pseudo-header, and if we find that the header object already has such a header, then we throw a stream protocol error.
Also added a test for this.

Result:
Header validation now rejects duplicate pseudo-headers, in line with best practices and the draft spec.