Add check point to methods `FastThreadLocalThread.threadLocalMap()` and `FastThreadLocalThread.setThreadLocalMap(...)` (#12057)
Motivation:

`FastThreadLocalThread.threadLocalMap()` and `FastThreadLocalThread.setThreadLocalMap(...)` should firstly check whether the target thread is the current thread, and print warn log if not.
Actually, IMHO, we can throw `Exception` here, but for compatible consideration, it's might be better that we just print log.
OR, we can change the field `threadLocalMap` to `volatile threadLocalMap`, which I think it's not good because it will hurts performance.

Modification:

Add check point of whether the target thread is the current thread to methods `FastThreadLocalThread.threadLocalMap()` and `FastThreadLocalThread.setThreadLocalMap(...)`, and print warn log if not.

Result:

Prints warn log if the target thread is not the current thread.

Co-authored-by: laosijikaichele <laosijikaichele>