Use `Objects#requireNonNull` #12098)
Motivation:
Currently, we rely upon `ObjectUtil#checkNotNull` for checking if an object is not `null`. This method existed because Java 6 or 7 didn't have `Objects#requireNonNull`. Now Netty 5 is based on Java so we can use ``Objects#requireNonNull`.

Modification:
Removed and changed `ObjectUtil#checkNotNull` to `Objects#requireNonNull`

Result:
Use standard code for `null` checks.