Implement the new Buffer API in terms of the ByteBuf API
Motivation:
This will be useful for bridging the gap between the two APIs.
Integrators will be able to use this as they migrate their code to the new API bit-by-bit.

Modification:
Add a complete implementation of the Buffer interface, that is written in terms of the ByteBuf API.
The implementation passes all of the Buffer tests.

Result:
It is now possible to adapt a ByteBuf into a Buffer.
This capability joins the ByteBufAdaptor as a sibling.
The ByteBufAdaptor converts a Buffer into a ByteBuf, while the new ByteBufBuffer works in the opposite direction, and converts a ByteBuf into a Buffer.