Allow attaching additional metadata to ResourceLeakTrackers on creation (#12091)
Motivation:
Attaching metadata to the creation record of a ResourceLeakTracker allows implementations to give more context information about where the leak originated from, such as the currently running test case, or, for fuzzing tests, the input that caused the leak.

Modification:
Add a protected method getInitialHint to ResourceLeakDetector that is called to populate the hint of the first TraceRecord of a leak. The method returns null by default, keeping the old behavior.

Result:
Subclasses can override the new method to supply metadata that is then interpreted the same way as a hint passed to ResourceLeakTracker.track, and attached to the "Created at:" record.

Co-authored-by: Chris Vest <christianvest_hansen@apple.com>