Include .hprof files when uploading artifacts
Motivation:
These are created when a test runs out of heap memory, and we need them to be able to effectively debug OOMEs in our builds.

Modification:
Make the upload-artifact action include any hprof files anywhere, when it wraps up the state of a failed build.

Result:
We should now get .hprof files included in the archives when a build fails with an OutOfMemoryError.