Change the default pooled byte buf allocator max order from 11 to 9
Motivation:
A max order of 11 implies a chunk size of 16 MiB, which can be a large investment in memory, and occasionally causes the tests to run out of memory.
It is also a large investment on machines with many cores, since the number of arenas are correlated with the number of CPU cores, and each arena that is used will likely allocate a chunk.

Modification:
Reduce the default max order such that the default chunk size is reduced from 16 MiB to 4 MiB.
This potentially increases the number of huge allocations, but I think it's a fair trade-off for Netty's use cases.

Result:
If we don't need a lot of memory, we'll now use less memory.