Reduce the default PooledByteBufAllocator chunk size from 16 MiB to 4 MiB
Motivation:
By default we allocate 2 arenas per core, and each arena that is put to use will allocate a chunk.
If we don't need a lot of memory, and certainly not compared to the number of cores on a system, then this will take up more memory than necessary, since each chunk is 16 MiB.
By reducing the chunk size to 4 MiB, we reduce the minimum memory usage by a good deal, in these cases where not much is needed.
The drawback is that we risk allocating more huge buffers, but this is a fair trade-off since Netty's use cases mostly involve very small buffers.

Modification:
Reduce the default max order from 11 to 9.
Also make similar configuration changes in PooledByteBufAllocatorTest, to reduce the memory usage during testing.

Result:
Netty now uses less memory when less memory is needed by the application.
This fixes #8536