Add PlatformDependent.estimateMaxDirectMemory
Motivation:
Integrators may want to also have a cross-platform way to compute the max direct memory.
Since we do this already, we can just as well expose the method.
However, all methods in PlatformDependent remain internal and unsupported, so callers should be robust to changing signatures and method names.

Modification:
Rename the maxDirectMemory0 method to estimateMaxDirectMemory and make it public.

Result:
Fixes #12103