Add BufferConversionHandler
Motivation:
Migrating code from one buffer API to the next will require a transition period.
For this purpose, it is useful to have a handler that can convert between the buffer types, so new and old handlers alike can participate in the same pipeline.

Modification:
Add a BufferConversionHandler that can be configured to convert reads, writes, and user events, between ByteBuf and Buffer.
Messages that are not either ByteBufs or Buffers will be passed through unchanged.

Result:
It is now easier to build pipelines that make use of both kinds of buffers.
Note that higher-level messages, such as HttpRequest or ByteBufHolder implementers, cannot be automatically converted in this way.
These objects pass through this handler unchanged.