Preserver io.netty group id but change artifact id
Motivation:
Changing the group id part of the Maven coordinates require us to have control over the netty5.io domain.
One domain might be fine, but it's clear that this will not scale in the future, for Netty 6, 7, etc.

Modification:
Instead of changing the group id, we annotate the artifact id, so e.g. `netty-common` becomes `netty5-common`.

Result:
We preserve our `io.netty` group id, while still allowing Netty 5 and 4 to co-exist as dependencies, and on the classpath.