Add ByteBufConvertible to Netty 4.1
Motivation:
This interface is used in the main branch to bridge buffer implementations.
Moving it to 4.1 will allow Netty 5 to use the 4.1 ByteBuf implementation.

Modification:
Copy the ByteBufConvertible interface from the main branch to 4.1.
And make ByteBuf implement the interface in the same way it does in the main branch.

Result:
The current main branch code can theoretically be made to use the ByteBuf implementation from 4.1.