Fix a race condition in DefaultChannelPipelineTest
Motivation:
Handles are added on the event loop thread, and external calls will be submitted to the event loop thread as tasks.
The test did not take any steps to ensure that such tasks had completed, before making assertions about the state of the pipeline.

Modification:
Add a count down latch as a barrier between the handler added callbacks, and the test making assetions about the state of the pipeline.

Result:
One fewer flaky tests.