Remove Android 5.0 work-around in SslHandler
Motivation:
The referenced bug was fixed in 2015, 3 years prior to the release of Java 11.
Thus, I think we are safe to assume that it's fixed in versions of Android that are relevant to Netty 5.

Modification:
Remove work-around for an Android 5.0 bug that has been fixed for a while.

Result:
Cleaner code.