Fixes that allow the aggregate javadocs to be generated (#12185)
Motivation:
Generating javadocs in a post-9 world means we need to be more strict about modules.

Modification:
Add automatic module names to modules that were missing these.

Add a couple of JDK modules to the javadoc module path.
Automatic modules automatically reads from all modules on the module path, so these need to be present for our modules to be able to see e.g. `java.naming` and `jdk.unsupported`.

Result:
It is now possible to generate javadocs, though we still get tons of errors reported, due to not living up to the standard required for Java 11.
The javadoc tool still generates javadocs for us, though.