Allow aliasing between read-only buffers
Motivation:
Aliasing is only a problem when mutation is involved.
Since making a buffer read-only is irreversible, we can permit them to share the underlying memory.

Modification:
Add a `Buffer.copy` method that takes an argument for whether the resulting buffer should be read-only or not.
If both original and copy buffers are read-only, then reuse the existing structural sharing mechanics of the const buffers, to make both read-only buffers share the same underlying memory.

Result:
Copying read-only buffers can now offer a useful optimisation.