Make it possible to tell if an allocation type is direct
Motivation:
This is the main use case for inspecting the allocation type of any buffer allocator.
We can simplify this and increase our confidence that it is correct by adding a getter for this property.

Modification:
Add an `isDirect` method on `AllocationType`.

Result:
It is now easier to tell if a `BufferAllocator` wil allocate direct buffers or not.