Attach lifecycle traces to ByteBufAdaptor reference count errors
Motivation:
Problems with reference counts can be difficult to debug without a lifecycle trace, so it's important that we attach traces to these exceptions.

Modification:
Add a `Statics` method for extracting the trace of a buffer and attaching it to a given exception.
Then use this when `release` is called on a `ByteBufAdaptor` that has already been fully released.

Result:
Easier to debug reference counting errors with `ByteBufAdaptor`.