Fix composite buffer invariant escape
Motivation:
It was possible to construct a composite buffer that had its internal invariants violated.
Specifically, gaps could be created in the readable and writable byte regions.

Modification:
Fix the gap checking code, which run every time we modify the component structure of a composite buffer.

Result:
It should no longer be possible to create a composite buffer that is internally inconsistent and broken in weird ways.