Add a SensitiveBufferAllocator
Motivation:
We have use cases in PemPrivateKey and sensitive PemValue objects, for buffers that can zero out themselves when closed.
Users may have use cases for sensitive buffers as well.

Modification:
Add a SensitiveBufferAllocator that functions similarly to the unpooled off-heap allocator, except the buffers are equipped with a drop implementation that zero out the memory of the buffer when it is deallocated.
The sensitive buffers otherwise behave exactly the same as any other off-heap buffer.
The allocator is not pooling because we don't anticipate a high allocation rate for sensitive buffers.
The sensitive buffers only come in an off-heap variant, since a GC may freely produce copies of any on-heap memory, eliminating our control of the lifetime of the data.
In the process of adding this, some unnecessary usages of ArcDrop have also been removed to aid testing, and incidentally also reduces overhead.

Result:
It is possible to reliably allocate buffers that zero out their memory when deallocated, thus reducing (but not eliminating) the window for memory scraping attacks.