Fix bug in read-only buffer copy
Motivation:
When a read-only copy of a read-only buffer was requested, we would rely on the infrastructure for const buffers to create structurally shared buffers and avoid actually copying.
Unfortunately, we forgot to return this structurally shared buffer copy in the `NioBuffer` and `UnsafeBuffer` implementations, and instead fell through to the traditional copying code.
The effect of this is that we leaked the structurally shared copy, and also did the work of actually copying the buffer.

Modification:
Return the structurally sharing copy - if we make one - from the copy methods of `NioBuffer` and `UnsafeBuffer`.

Result:
No leak, and intended optimisation actually realised.