Fix a regression that caused the DirectByteBuffer constructor to be unavailable
Motivation:
The check for the availability had been modified to account for changes in the constructor arguments in Java 17.
Unfortunately, these changes broke the check for all versions, and made us fall back to plain ByteBuffer-based implementations.

Modification:
Fix the check. Specifically, we cannot use `invokeExact` because the constructor returns a `DirectByteBuffer` object, and that class is package-private to `java.nio`.

Result:
We can now use unsafe buffer implementations again.