Release our strong references to memory faster
Motivation:
Now that PoolArenas are once again strongly referenced by the PoolThreadCache, it turns out long-lived threads can keep a lot of memory alive from closed pooled allocators.

Modification:
When the PooledBufferAllocator is closed, we go out of our way to null-out references to PoolChunks in the smallSubpagePools array, and in the PoolChunkLists.
Loosing the references to the closed PoolChunks allows the GC to collect the heap memory that might have been held alive by the PoolChunk.memory field.

Result:
We no longer accidentally hold on to memory we don't need.