Make `BufferLeakDetectionTest` print less to the log
Motivation:
The `BufferLeakDetectionTest` was set up to log warnings about leaks that were not part of the test.
In one of the tests, this mechanism backfired on itself, and printed lots of log output.
The test was installing two leak detectors - using two different hints - and checking that only some objects were detected as leaked.
However, the leak of the intended type also triggered the detector for the unintended type, and made it log a warning.

Modification:
Make it configurable in the test, if warnings should be printed, and then silence the warnings in the case where we know the detector will find something that doesn't match its hint.

Result:
Much less log output from `BufferLeakDetectionTest`.