Cope with misbehaving `Http2Headers` implementations
Motivation:
Some integrators may choose to use custom `Http2Headers` implementations, and might not implement all methods, or might not anticipate the ways in which we'll use the methods.
For instance, this was a problem that affected GRPC.

Modification:
To cope with this, we guard our newly added call to `Http2Headers.contains()` call with an `instanceof` check to see if we're familiar with the implementation, and fall back to checking the direct accessor methods if not.

Result:
Avoid breaking those who wish to use old versions of GRPC with newer versions of Netty, or situations like that.