Log specific exception to debug ParameterizedSslHandlerTest
Motivation:
We have been seeing this test fail intermittently on Windows for some time.
The stack trace of the exception was supposed to be included in the test failure output, but we don't see it in practice.

Modification:
Add some logging to the test, so when it is about to fail, it will capture additional context in the log.

Result:
Hopefully we'll be able to debug this test in the future.