Fix data race bug in `PooledBufferAllocator`
Motivation:
When it comes to the `ByteBuffer` based Buffer implementation, the `MemoryManager.sliceMemory()` method works by mutating position and limit of the underlying `ByteBuffer` object, that in the case of `PoolChunk` is the shared memory field.
These mutations must be done under the protection of the locks in the pool, so the slicing must be done eagerly rather than lazily.

Modification:
Move the `sliceMemory()` call from the `memory()` accessor method to the constructor of `UntetheredChunkAllocation`.
The constructor is called under the protection of the locks inside the allocator, while the `memory()` accessor was not.

Result:
The `PooledBufferAllocator` should now be as thread-safe as it was intended to be.