Keep buffers alive for the duration of forEach calls
Motivation:
There was a theoretical issue where a Buffer instance could be garbage collected while the ByteBuffers from forEach components were still in use.
The issue exist because the Buffer instance itself is not necessarily used past the calls to acquiring ByteBuffers, and thus the compiler and GC can conspire to collect the Buffer instance, which in turn can lead the Cleaner to free the underlying memory that is still used by the ByteBuffer.

Modification:
Enclose the component processing in NioBuffer and UnsafeBuffer with try-finally clauses that adds a reachability fence to keep the buffers alive throughout the component processing.

Result:
Buffers can no longer be freed while component processing inside a forEach method is taking place.