Make maven module build order more clear on Windows (#12308)
Motivation:
On Windows, some Maven versions on might not identify that `netty-testsuite-native` needs to have the classes for our native modules compiled first.
This can make the build fail with a compilation error saying that e.g. the epoll or kqueue packages can't be found.

Modification:
The `netty-testsuite-native` now unconditionally depends on the classes of the native modules.
Then relies on the profiles for pulling in the native dependencies, like before.

Result:
Clean whole-project builds on Windows should no longer have any problems with compiling `netty-testsuite-native`.