Migrate DNS to Buffer
Motivation:
We wish to migrate to the new Buffer API and remove ByteBuf.
To that end, we will have to migrate our DNS implementation to the new API.

Modification:
Change all usages of ByteBuf to Buffer in the DNS code, and change usages of DatagramPacket to BufferDatagramPacket.

Result:
Our DNS codec and resolver implementation now use the Buffer API.