Enable debugging of asynchronous tasks in Intellij
Motivation:
Evented systems like Netty can be much harder to debug than sequential and blocking code.
Fortunately, with a few strategically placed hints, we can significantly improve the tooling support.

Modification:
Add a dependency on the JetBrains annotations, and use the `Schedule` and `Execute` annotations to tell the Intellij IDEA debugger how our event loops and executors handle asynchronous task execution.
Basically, we tell it where tasks are scheduled for asynchronous execution, and we tell it where this execution is picked up.

Result:
The Intellij IDEA debugger can then bridge the schedule and execute points, and present a unified, or spliced stack trace.
This way, we are able to see who scheduled what tasks, and why.