Fix compilation on Mac OS
Motivation:
My compiler require that the const-ness of the package prefix is preserved.

Modification:
Make the `staticPackagePrefix` variable `const`.
The `const` qualifier here means that the value pointed to is unassignable.
However, we are still allowed to assign to the pointer (the pointer itself isn't `const`), which is why the rest of the code does not complain about this change.

Result:
No more native compilation errors on MacOS.