Make `Buffer.write*` methods automatically
Motivation:
Depending on the format, it can be hard to allocate correctly sized buffers up-front.
In these cases, it is quite convenient if the primitive write methods automatically expand the buffer until some threshold.

Modification:
Add logic to the bounds checking to automatically expand buffers when used by a primitive write method.

Result:
Buffers are now more convenient to use for writing, when the amount of data cannot be easily determined a priori.