Prototype of an external component iteration API
Motivation:
Accessing the internal ByteBuffer instances, or the native addresses, with the `Buffer.forEachReadable/Writable` methods have turned out to be rather troublesome.
The trouble specifically comes from the API being designed around _internal_ iteration.
While internal iteration is excellent at masking whether a buffer is composite or not, the loss of local mutable context (because of the use of lambdas) makes it rather combersome to use.

Modification:
This adds a prototype of an API based on external iteration of the components.
There are a number of requirements pulling on this API and causing it to look the way it does.
First, we really want to ensure that the parent `Buffer` instance is strongly referenced and live throughout the entire iteration.
With internal iteration, we can place a `reachabilityFence` after the iteration.
With external iteration, we instead rely on the `ComponentIterator` being `AutoCloseable` and place a reachability fence in the close method.
Second, we wish to prevent the components from leaking beyond the iteration.
This isn't entirely possible to prevent, for either internal or external iteration.
The use of intersection types discourages assigning the iterator or the components to fields, though they don't prevent it.
It does, however, prevent correctly typed local variables from existing beyond their intended scope.

Result:
We have a concrete API proposal for external buffer component iteration that we can talk about.