Make it thread-safe to call MemoryManager#availableManagers
Motivation:
It turns out ServiceLoader#stream is not thread-safe, so we cannot cache
the service loader directly.

Modification:
Create a lazily populated cache based on an immutable List, and protect
the ServiceLoader stream with a lock.
The returned stream now becomes eager, because the backing immutable
list is populated fully before streamed.
This is probably not a problem in practice, but theoretically, if
a provider throws exceptions from their static initialisers, it will now
prevent the stream from fully populating, and thus the cache from being
created.

Result:
We can now call MemoryManager#availableManagers in parallel.