Run the netty-buffer tests in parallel
Motivation:
We have a lot of test cases to go through, due to the number of combinations and buffer states explored by the parameterised tests.

Modification:
Tell the JUnit 5 runner platform to run the buffer tests in parallel, per class.
Within each class, the tests will run sequentially.
This preserves the existing setup/tear-down behaviour.
The BufferCleanerTest is an exception, due to how it accesses Statics.MEM_USAGE_NATIVE, so it runs in isolation, when no other tests are running.

Result:
Faster build.