Make BufferConversionHandler(Conversion) constructor more logical
Motivation:
It's natural to expect the conversion on writes to be the inverse of reads.

Modification:
Add an invert method to Conversion, and make it so that when you only specify one Conversiont to the BufferConversionHandler constructor, the write-conversion is automatically the inverse of the read-conversion.

Result:
Conversion handler behaviour is less surprising.